# schemas/ripd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RIP Agent Server",
    "description": "Agent session control plane (HTTP/SSE).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/threads/{id}": {
      "get": {
        "operationId": "thread_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadMeta"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/branch": {
      "post": {
        "operationId": "thread_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Parent thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadBranchPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadBranchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid branch request"
          },
          "404": {
            "description": "Thread or branch point not found"
          }
        }
      }
    },
...
    "/threads/{id}/context-selection-status": {
      "post": {
        "operationId": "thread_context_selection_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextSelectionStatusV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Context selection strategy decisions (truth-derived)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextSelectionStatusV1Response"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/events": {
      "get": {
        "operationId": "thread_stream_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of thread event frames"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/handoff": {
      "post": {
        "operationId": "thread_handoff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadHandoffPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Handoff thread created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadHandoffResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid handoff request"
          },
          "404": {
            "description": "Thread or handoff point not found"
          }
        }
      }
    },
    "/threads/{id}/messages": {
      "post": {
        "operationId": "thread_post_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadPostMessagePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Message accepted and run started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadPostMessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/provider-cursor-rotate": {
      "post": {
        "operationId": "thread_provider_cursor_rotate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCursorRotateV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Provider cursor rotation logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCursorRotateV1Response"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/provider-cursor-status": {
      "post": {
        "operationId": "thread_provider_cursor_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCursorStatusV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Provider cursor status projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCursorStatusV1Response"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContextSelectionStatusCheckpointV1": {
        "type": "object",
        "required": [
          "checkpoint_id",
          "summary_kind",
          "summary_artifact_id",
          "to_seq"
        ],
        "properties": {
          "checkpoint_id": {
            "type": "string"
          },
          "summary_artifact_id": {
            "type": "string"
          },
          "summary_kind": {
            "type": "string"
          },
          "to_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ContextSelectionStatusDecisionV1": {
        "type": "object",
        "required": [
          "decision_event_id",
          "run_session_id",
          "message_id",
          "compiler_id",
          "compiler_strategy",
          "limits",
          "resets",
          "actor_id",
          "origin",
          "seq",
          "timestamp_ms"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "compaction_checkpoint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContextSelectionStatusCheckpointV1"
              }
            ]
          },
          "compaction_checkpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextSelectionStatusCheckpointV1"
            }
          },
          "compiler_id": {
            "type": "string"
          },
          "compiler_strategy": {
            "type": "string"
          },
          "decision_event_id": {
            "type": "string"
          },
          "limits": {},
          "message_id": {
            "type": "string"
          },
          "origin": {
            "type": "string"
          },
          "reason": {},
          "resets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextSelectionStatusResetV1"
            }
          },
          "run_session_id": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "timestamp_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ContextSelectionStatusResetV1": {
        "type": "object",
        "required": [
          "input",
          "action",
          "reason"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "input": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "ref": {}
        }
      },
      "ContextSelectionStatusV1Request": {
        "type": "object",
        "properties": {
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ContextSelectionStatusV1Response": {
        "type": "object",
        "required": [
          "thread_id",
          "decisions"
        ],
        "properties": {
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextSelectionStatusDecisionV1"
            }
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
...
      "ProviderCursorRotateV1Request": {
        "type": "object",
        "required": [
          "actor_id",
          "origin"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin": {
            "type": "string"
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ProviderCursorRotateV1Response": {
        "type": "object",
        "required": [
          "thread_id",
          "rotated"
        ],
        "properties": {
          "cursor_event_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "rotated": {
            "type": "boolean"
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
      "ProviderCursorStatusCursorV1": {
        "type": "object",
        "required": [
          "cursor_event_id",
          "provider",
          "action",
          "actor_id",
          "origin",
          "seq",
          "timestamp_ms"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "actor_id": {
            "type": "string"
          },
          "cursor": {},
          "cursor_event_id": {
            "type": "string"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "run_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "timestamp_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ProviderCursorStatusV1Request": {
        "type": "object"
      },
      "ProviderCursorStatusV1Response": {
        "type": "object",
        "required": [
          "thread_id",
          "cursors"
        ],
        "properties": {
          "active": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProviderCursorStatusCursorV1"
              }
            ]
          },
          "cursors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderCursorStatusCursorV1"
            }
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
...
      "ThreadBranchPayload": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadBranchResponse": {
        "type": "object",
        "required": [
          "thread_id",
          "parent_thread_id",
          "parent_seq"
        ],
        "properties": {
          "parent_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "parent_thread_id": {
            "type": "string"
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
...
      "ThreadHandoffPayload": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary_artifact_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary_markdown": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadHandoffResponse": {
        "type": "object",
        "required": [
          "thread_id",
          "from_thread_id",
          "from_seq"
        ],
        "properties": {
          "from_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "from_thread_id": {
            "type": "string"
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
      "ThreadMeta": {
        "type": "object",
        "required": [
          "thread_id",
          "created_at_ms",
          "archived"
        ],
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "created_at_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "thread_id": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadOpenResponsesOverrides": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "followup_user_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "parallel_tool_calls": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "reasoning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OpenResponsesReasoningConfig"
              }
            ]
          },
          "stateless_history": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
      "ThreadPostMessagePayload": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "actor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": "string"
          },
          "openresponses": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ThreadOpenResponsesOverrides"
              }
            ]
          },
          "origin": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadPostMessageResponse": {
        "type": "object",
        "required": [
          "thread_id",
          "message_id",
          "session_id"
        ],
        "properties": {
          "message_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}